+++ src/Interfaces/Values/AlfEmptyGet.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Alf\Interfaces\Values;
+
+use JetBrains\PhpStorm\Pure;
+
+interface AlfEmptyGet extends AlfValueGet {
+
+    #[Pure]
+    public function isEmpty() : bool;
+
+    #[Pure]
+    public function getEmptyValue();
+
+}