+++ src/Interfaces/Values/AlfEmptySet.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Alf\Interfaces\Values;
+
+interface AlfEmptySet {
+
+    public function setToEmpty() : static;
+
+}