+++ src/Interfaces/Values/AlfNullGet.php
@@ -2,8 +2,11 @@
 
 namespace Alf\Interfaces\Values;
 
+use JetBrains\PhpStorm\Pure;
+
 interface AlfNullGet extends AlfValueGet {
 
+    #[Pure]
     public function isNull() : bool;
 
 }