+++ src/Interfaces/Values/AlfValueGet.php
@@ -2,8 +2,11 @@
 
 namespace Alf\Interfaces\Values;
 
+use JetBrains\PhpStorm\Pure;
+
 interface AlfValueGet {
 
+    #[Pure]
     public function getValue();
 
 }