+++ src/Types/Scalars/AlfInt.php
@@ -6,6 +6,7 @@ use Alf\AlfBasicTypeScalar;
 use Alf\Attributes\AlfAttrAutoComplete;
 use Alf\Attributes\AlfAttrParameterIsInt;
 use Alf\Services\AlfProgramming;
+use JetBrains\PhpStorm\ArrayShape;
 use JetBrains\PhpStorm\Pure;
 
 class AlfInt extends AlfBasicTypeScalar {
@@ -13,7 +14,7 @@ class AlfInt extends AlfBasicTypeScalar {
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
     final public static function _AlfInt($obj) : AlfInt {
-        return AlfProgramming::_()->unused($obj, static::_AlfBasicTypeScalar($obj), static::_AlfNullGet($obj));
+        return AlfProgramming::_()->unused($obj, static::_AlfBasicTypeScalar($obj));
     }
 
     protected ?int $value = null;
@@ -26,15 +27,19 @@ class AlfInt extends AlfBasicTypeScalar {
     public function set(#[AlfAttrParameterIsInt] int|null|AlfInt $value) : static {
         $newValue = AlfProgramming::_()->valueToInt($value);
         if (is_null($newValue)) {
-            return $this->setToNull();
+            $this->value = null;
+            return $this;
         }
         $this->value = $this->_convertValueForSet($newValue);
         return $this;
     }
 
     public function setToNull() : static {
-        $this->value = null;
-        return $this;
+        return $this->set(null);
+    }
+
+    public function setToEmpty() : static {
+        return $this->set($this->getEmptyValue());
     }
 
     #[Pure]
@@ -50,6 +55,18 @@ class AlfInt extends AlfBasicTypeScalar {
     #[Pure]
     public function getValue() : ?int {
         return $this->value;
+    }
+
+    #[Pure]
+    public function getEmptyValue() : int {
+        return 0;
+    }
+
+    #[ArrayShape(['class' => "string", 'parent' => "string", 'value' => "int|string"])]
+    public function __debugInfo() : array {
+        $output = parent::__debugInfo();
+        $output['value'] = ($this->isNull() ? '-NULL-' : $this->getValue());
+        return $output;
     }
 
 }